refactor: move utils folder
