--- src/app/utils/constants/proyectoFields.ts
@@ -1,51 +1,0 @@
-export const proyectoFields = [
-  {
-    id: "codigo",
-    label: "Codigo",
-    type: "text",
-    required: true,
-    placeholder: "Introduce el codigo para el proyecto...",
-  },
-  {
-    id: "ip",
-    label: "Investigador Principal",
-    type: "text",
-    required: true,
-    placeholder: "Nombre del investigador principal...",
-  },
-  {
-    id: "coip",
-    label: "Co-Investigador Principal",
-    type: "text",
-    required: false,
-    placeholder: "Nombre de autor del co-investigador principal...",
-  },
-  {
-    id: "titulo",
-    label: "Titulo",
-    type: "text",
-    required: true,
-    placeholder: "Titulo del proyecto...",
-  },
-  {
-    id: "financiado",
-    label: "Entidad financiadora",
-    type: "text",
-    required: true,
-    placeholder: "Entidad financiadora...",
-  },
-  {
-    id: "inicio",
-    label: "Fecha de inicio del proyecto",
-    type: "date",
-    required: true,
-    placeholder: "Introduce la fecha de inicio...",
-  },
-  {
-    id: "fin",
-    label: "Fecha de finalización del proyecto (opcional)",
-    type: "date",
-    required: false,
-    placeholder: "Introduce la fecha de fin...",
-  },
-]
